requirements.txt
Flask==1.1.1
tensorflow==2.3.1
Pillow==8.0.0
regex==2020.4.4
numpy==1.18.5
pandas==1.1.3
matplotlib==3.2.1
keras==2.4.3
opencv-python==4.4.0.46
annoy==1.17.0
pyaml==20.4.0
ruamel-yaml==0.15.80

